requirements.txt
python-telegram-bot==21.4
gspread==5.12.0
oauth2client==4.1.3
waitress==2.1.2
flask==2.0.3
python-dotenv==0.19.2